project/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2==2.9.10
PyJWT==2.10.1
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.1
